fix build on msvc2019, add example

+++ src/random.h
@@ -51,7 +51,7 @@ std::array<uint8_t, N> random_array() {
   std::random_device rd;
   std::mt19937 engine(rd());
 
-  std::uniform_int_distribution<uint8_t> uniform_dist(0, 255);
+  std::uniform_int_distribution<unsigned int> uniform_dist(0, 255);
 
   // Fill block with random values.
   std::array<uint8_t, N> array;
